(PC-7905): Move AllocineVenueProviderRule to subpackage core.provider
