+++ tests/use_cases/connect_allocine_to_venue_test.py
@@ -3,8 +3,8 @@ from decimal import Decimal
 from pcapi.core.offers.factories import VenueFactory
 from pcapi.core.providers.factories import AllocinePivotFactory
 from pcapi.core.providers.models import AllocineVenueProvider
+from pcapi.core.providers.models import AllocineVenueProviderPriceRule
 from pcapi.model_creators.provider_creators import activate_provider
-from pcapi.models import AllocineVenueProviderPriceRule
 from pcapi.routes.serialization.venue_provider_serialize import PostVenueProviderBody
 from pcapi.use_cases.connect_venue_to_allocine import connect_venue_to_allocine
 from pcapi.utils.human_ids import humanize
